+++ brightbox/provider.go
@@ -73,6 +73,7 @@ func Provider() terraform.ResourceProvider {
 			"brightbox_load_balancer":   resourceBrightboxLoadBalancer(),
 			"brightbox_database_server": resourceBrightboxDatabaseServer(),
 			"brightbox_container":       resourceBrightboxContainer(),
+			"brightbox_api_client":      resourceBrightboxApiClient(),
 		},
 		ConfigureFunc: providerConfigure,
 	}