+++ brightbox/resource_brightbox_firewall_policy.go
@@ -70,6 +70,11 @@ func resourceBrightboxFirewallPolicyRead(
 
 	firewall_policy, err := client.FirewallPolicy(d.Id())
 	if err != nil {
+		if strings.HasPrefix(err.Error(), "missing_resource:") {
+			log.Printf("[WARN] Firewall Policy not found, removing from state: %s", d.Id())
+			d.SetId("")
+			return nil
+		}
 		return fmt.Errorf("Error retrieving Firewall Policy details: %s", err)
 	}