+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add timeout support for database servers
 - Add timeout support for servers
 - Add timeout support for Cloud IPs
+- Add API client resource
 
 ## 1.0.6 (March 04, 2019)
 - Switch to Go Modules #3